requirements.txt
torch
transformers
diffusers
accelerate
pillow
vllm
sglang
